requirements.txt
Django==5.2.*
psycopg2-binary==2.9.*
gunicorn==22.*
django-htmx==1.*
django-auditlog==3.*
django-simple-history==3.*
django-two-factor-auth==1.*
django-otp==1.*
django-axes==6.*
python-magic==0.4.*
Pillow==10.*
django-environ==0.11.*
whitenoise==6.*
